+++ src/main/java/com/alkemy/ong/data/gateways/DefaultOrganizationGateway.java
@@ -1,11 +1,11 @@
 package com.alkemy.ong.data.gateways;
 
 import com.alkemy.ong.data.entities.OrganizationEntity;
-import com.alkemy.ong.data.mapper.OrganizationModelMapper;
+
 import com.alkemy.ong.data.repositories.OrganizationRepository;
 import com.alkemy.ong.domain.organizations.Organization;
 import com.alkemy.ong.domain.organizations.OrganizationGateway;
-import net.bytebuddy.implementation.bytecode.Throw;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -19,8 +19,6 @@ public class DefaultOrganizationGateway implements OrganizationGateway {
     @Autowired
     private OrganizationRepository organizationRepository;
 
-    @Autowired
-    private OrganizationModelMapper organizationModelMapper;
 
     @Override
     public Organization findById(Long id){
@@ -28,9 +26,27 @@ public class DefaultOrganizationGateway implements OrganizationGateway {
         if(!organizationEntity.isPresent()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-        Organization organization= organizationModelMapper.entity2Model(organizationEntity.get());
+        Organization organization= entity2Model(organizationEntity.get());
         return organization;
     }
 
+    public OrganizationEntity model2Entity(Organization organization){
+        return OrganizationEntity.builder()
+                .address(organization.getAddress())
+                .name(organization.getName())
+                .image(organization.getImage())
+                .phone(organization.getPhone())
+                .build();
+
+    }
+
+    public Organization entity2Model(OrganizationEntity organizationEntity){
+        return Organization.builder()
+                .name(organizationEntity.getName())
+                .phone(organizationEntity.getPhone())
+                .image(organizationEntity.getImage())
+                .address(organizationEntity.getAddress())
+                .build();
+    }
 
 }
